requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
psycopg2>=2.9.6  
redis>=4.6.0  
django-redis>=5.3.0  
pytest>=7.4.0  
pytest-django>=4.5.2  
drf-yasg>=1.21.5  
gunicorn>=21.2.0 
